refactor(ai-gateway): Inline the MCP announcement into the chat endpoint
The announcement is three statements and one constant, and its only caller is the
chat endpoint's own announceMCP, so it lives there as a method rather than in a
separate file behind a second function.

Folding the two together also drops the mcpServer field that announceMCP used as its
gate. NewHandler set that field and mcpBaseURL together, and only when the
turn-scoped endpoint was mounted, so a nil server and an empty base URL were the same
condition; mcpBaseURL alone now carries it. That takes the MCP SDK out of the chat
endpoint entirely — the gateway no longer reaches through the turn-scoped endpoint
for a server it only tested for nil.
