+++ cmd/jaeger/internal/extension/jaegerquery/internal/jaegerai/endpoint_chat_test.go
@@ -18,7 +18,6 @@ import (
 	aguitypes "github.com/ag-ui-protocol/ag-ui/sdks/community/go/pkg/core/types"
 	"github.com/coder/acp-go-sdk"
 	"github.com/gorilla/websocket"
-	"github.com/modelcontextprotocol/go-sdk/mcp"
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
 	"go.opentelemetry.io/otel/attribute"
@@ -407,6 +406,57 @@ func TestChatEndpointRegistersTurnInRegistry(t *testing.T) {
 	assert.Equal(t, 0, turns.count(), "the stream must be removed at end of turn")
 }
 
+func httpCaps(supported bool) acp.AgentCapabilities {
+	return acp.AgentCapabilities{
+		McpCapabilities: acp.McpCapabilities{Http: supported},
+	}
+}
+
+const testBaseURL = "https://jaeger.example.com:16686"
+
+// announceEndpoint is a chatEndpoint reduced to the fields announceMCP reads, so the
+// announcement can be exercised without standing up a sidecar.
+func announceEndpoint(baseURL, basePath string) *chatEndpoint {
+	return &chatEndpoint{mcpBaseURL: baseURL, basePath: basePath}
+}
+
+func TestAnnounceMCPOverHTTP(t *testing.T) {
+	got := announceEndpoint(testBaseURL, "").announceMCP(httpCaps(true), "route-1")
+	require.Len(t, got, 1)
+	require.NotNil(t, got[0].Http)
+	assert.Equal(t, "http", got[0].Http.Type)
+	assert.Equal(t, mcpServerName, got[0].Http.Name)
+	assert.Equal(t, testBaseURL+"/api/ai/mcp/route-1/", got[0].Http.Url)
+}
+
+// TestAnnounceMCPRequiresHTTPCapability is the core contract: ACP requires an agent
+// to opt into each McpServer variant, so announcing HTTP to an agent that cannot
+// consume it would make it fail the session.
+func TestAnnounceMCPRequiresHTTPCapability(t *testing.T) {
+	got := announceEndpoint(testBaseURL, "").announceMCP(httpCaps(false), "route-1")
+	assert.Empty(t, got, "an agent that did not advertise mcpCapabilities.http is offered nothing")
+}
+
+// TestAnnounceMCPRequiresBaseURL covers both reasons mcpBaseURL is empty: the
+// turn-scoped endpoint is not mounted at all, or it is mounted but no reachable
+// address is known (see AIConfig.resolveMCPBaseURL).
+func TestAnnounceMCPRequiresBaseURL(t *testing.T) {
+	got := announceEndpoint("", "").announceMCP(httpCaps(true), "route-1")
+	assert.Empty(t, got, "with no reachable base URL the endpoint is not announced")
+}
+
+func TestAnnounceMCPRequiresRouteID(t *testing.T) {
+	got := announceEndpoint(testBaseURL, "").announceMCP(httpCaps(true), "")
+	assert.Empty(t, got, "without a route id there is no turn-scoped endpoint to announce")
+}
+
+func TestAnnounceMCPEmbedsBasePath(t *testing.T) {
+	got := announceEndpoint("http://127.0.0.1:16686", "/jaeger").announceMCP(httpCaps(true), "u-1")
+	require.Len(t, got, 1)
+	assert.Equal(t, "http://127.0.0.1:16686/jaeger/api/ai/mcp/u-1/", got[0].Http.Url,
+		"the announced URL must carry the query server's base path")
+}
+
 // TestChatEndpointAnnouncesMCPEndpoint is the end-to-end wiring check: the turn's
 // route id is registered in the turn registry and announced to the sidecar as a
 // reachable URL, so the agent can finally dial the endpoint that #8973 made serve
@@ -421,7 +471,6 @@ func TestChatEndpointAnnouncesMCPEndpoint(t *testing.T) {
 	defer cleanup()
 
 	handler := newChatEndpoint(zap.NewNop(), nil, newTurnRegistry(), wsURL, "/jaeger", 1<<20)
-	handler.mcpServer = mcp.NewServer(&mcp.Implementation{Name: "t", Version: "0"}, nil)
 	handler.mcpBaseURL = "https://jaeger.example.com:16686"
 
 	reqBody, err := json.Marshal(newAGUIRequest("hello"))
@@ -484,7 +533,6 @@ func TestChatEndpointAnnouncesNothingWithoutBaseURL(t *testing.T) {
 	defer cleanup()
 
 	handler := newChatEndpoint(zap.NewNop(), nil, newTurnRegistry(), wsURL, "", 1<<20)
-	handler.mcpServer = mcp.NewServer(&mcp.Implementation{Name: "t", Version: "0"}, nil)
 	// mcpBaseURL deliberately left empty — enable_mcp without ai.mcp_base_url.
 
 	reqBody, err := json.Marshal(newAGUIRequest("hello"))
